Change the manager tests into the TestCases pattern
